SNOW-1229108 Change interface naming convention
